Added default Auth initializer
